# twa/openapi/api.yml
openapi: 3.0.0
info:
  title: Posebestoimosti API
  description: API for managing deliveries and users
  version: 1.0.0
  contact:
    name: API Support
    email: support@posebestoimosti.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT    
servers:
  - url: http://prod.posebesto.ru:5001
    description: Production server
  - url: http://dev.posebesto.ru:5001
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /private/items/category/{category}:
    
    get:
      operationId: getItemsByCategory
      summary: Get items by category
      description: Retrieve a list of items for a specific category
      tags:
        - Items
      parameters:
        - name: category
          in: path
          required: true
          description: Category name
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: "OK"
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ItemOverview'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /private/customers/{customerId}/items/{itemId}:
    get:
      operationId: getCustomerItem
      summary: Get item overview for a customer
      description: Retrieve item overview for a specific customer and item
      tags:
        - Items
      parameters:
        - name: customerId
          in: path
          required: true
          description: Customer ID
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          description: Item ID
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: "OK"
                  data:
                    $ref: '#/components/schemas/ItemOverview'
        '404':
          description: Customer or item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /private/customers/{customerId}/cart/items:
    get:
      operationId: getCartItems
      summary: Get cart items for a customer
      description: Retrieve all items in a customer's cart
      tags:
        - Cart
      parameters:
        - name: customerId
          in: path
          required: true
          description: Customer ID
          schema:
            type: string
      responses:
        '200':
          description: Successful response with cart items
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: "OK"
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CartItem'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: addItemToCart
      summary: Add item to cart
      description: Add an item to a customer's cart
      tags:
        - Cart
      parameters:
        - name: customerId
          in: path
          required: true
          description: Customer ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToCartItem'
      responses:
        '200':
          description: Item successfully added to cart
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: "OK"
                  data:
                    $ref: '#/components/schemas/CartItem'
        '404':
          description: Customer or item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: removeItemFromCart
      summary: Remove item from cart
      description: Remove an item from a customer's cart
      tags:
        - Cart
      parameters:
        - name: customerId
          in: path
          required: true
          description: Customer ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoveFromCartItem'
      responses:
        '204':
          description: Item successfully removed from cart
        '404':
          description: Customer or cart item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /private/customers/{customerId}/cart/order:
    post:
      operationId: createOrderAndPaymentFromCart
      summary: Create order from cart and initialize payment
      description: Transactionally and idempotently fetch all cart items for the customer, split them by group, create separate orders for each group, find earliest delivery for each group, assign deliveries, initialize payment for each order, and return created orders and payments
      tags:
        - Cart
      parameters:
        - name: customerId
          in: path
          required: true
          description: Customer ID
          schema:
            type: string
        - name: idempotency_key
          in: header
          required: true
          description: Idempotency key to prevent duplicate order creation
          schema:
            type: string
            example: "unique-key-123"
      responses:
        '200':
          description: Orders created and payments initialized successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: "OK"
                  data:
                    type: object
                    properties:
                      orders:
                        type: array
                        items:
                          $ref: '#/components/schemas/Order'
                      payments:
                        type: array
                        items:
                          $ref: '#/components/schemas/Payment'
        '400':
          description: Bad request (e.g., cart is empty, invalid idempotency key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict (e.g., duplicate request with same idempotency key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /private/customers/{customerId}/subscriptions:
    post:
      operationId: createSubscription
      summary: Create a subscription for a customer
      description: Create a new subscription for a customer. If customer already has an active subscription which is not passed due (current date is before next_payment_date) throw bad request. If there is no active subscription for user create one and set next_payment_date month ahead.
      tags:
        - Subscriptions
      parameters:
        - name: customerId
          in: path
          required: true
          description: Customer ID
          schema:
            type: string
      responses:
        '200':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: "OK"
                  data:
                    type: object
                    properties:
                      paymentUrl:
                        type: string
                        description: URL for payment processing
                        example: "https://securepay.tinkoff.ru/payment/init?PaymentId=123456"        
        '400':
          description: Bad request (e.g., customer already has an active subscription)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /private/customers/{customerId}:
    get:
      operationId: getCustomerOverview
      summary: Get customer overview
      description: Retrieve customer overview including balance, stats, and subscription
      tags:
        - Customers
      parameters:
        - name: customerId
          in: path
          required: true
          description: Customer ID
          schema:
            type: string
      responses:
        '200':
          description: Successful response with customer overview
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: "OK"
                  data:
                    $ref: '#/components/schemas/CustomerOverview'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /private/customers/{customerId}/banks:
    get:
      operationId: getCustomerBanks
      summary: Get banks for a customer
      description: Retrieve a list of banks associated with a customer
      tags:
        - Customers
      parameters:
        - name: customerId
          in: path
          required: true
          description: Customer ID
          schema:
            type: string
      responses:
        '200':
          description: Successful response with list of banks
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: "OK"
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bank'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CustomerOverview: 
      type: object
      required:
        - balance
      allOf:
        - $ref: '#/components/schemas/Customer'
      properties:
        balance: 
          $ref: '#/components/schemas/CustomerBalance'
        stats:
          $ref: '#/components/schemas/CustomerStats'
        subscription:
          $ref: '#/components/schemas/Subscription'
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code
              example: "NOT_FOUND"
            message:
              type: string
              description: Error message
              example: "Delivery not found"
...
    AddToCartItem:
      type: object
      required:
        - itemId
      properties:
        itemId:
          type: string
          description: Item ID to add to cart
          example: "06088197-f7ef-4bf2-8b3f-2f4a05c6c104"
    RemoveFromCartItem:
      type: object
      properties:
        cartItemId:
          type: string
          description: Cart item ID to remove from cart
          example: "db13bdf1-cd12-4650-82be-f55af1e79bf7"
        itemId:
          type: string
          description:  item ID to remove from cart. All the cart items of this item will be removed
          example: "db13bdf1-cd12-4650-82be-f55af1e79bf7"
...